requirements.txt
gradio==4.44.1
pydantic==2.10.6
langchain
langchain_community
langsmith
chromadb
openai
tiktoken
pypdf
langchain-openai
python-dotenv